Rounded Container and FAB(s) can disappear sooner
For instances of where you want a ListView, GridView, or something similar to overlay perhaps a SliverAppBar, but with a rounded effect, now you can. (See [example post](https://stackoverflow.com/questions/54856511/))

It is now possible to have the FAB icons (or whatever Widget is placed inside of it) to disappear earlier by setting the disappearAt variable. Useful if you find it overlaying your Appbar, but you want it to disappear before it reaches your AppBar.